Consolidate action discovery into unified ActionZone component
Refactors the "Add Action" experience by replacing ActionPopover and
ActionSelectorNode with a single, multi-mode ActionZone component.

Key enhancements:
- Direct Operation Discovery: Users can find specific operations (e.g.
  'Clear', 'Email') without knowing the parent Action Type.
- Keyword Support: Added a keywords registry for alias-based discovery.
- On-the-fly Indexing: Integrated debounced backend fuzzy search for
  dynamic process operation discovery.
- UX Improvements: Fixed scrolling issues, added remote search spinner,
  and standardized keyboard navigation.
- Consolidation: Moved search logic into useActionSearch composable.
- Reliability: Fixed file path resolution and formatting for CI.

Includes audit_report.md documenting the architecture.
